aiq: release v2.0x60.0
sync to CL 661a387a

Details CL info:
661a387a (origin/isp2x) modify version to v2.0x60.0

+++ hwi/isp20/Isp20_module_dbg.h
@@ -20,6 +20,8 @@
 
 #define RUNTIME_MODULE_DEBUG
 #ifdef RUNTIME_MODULE_DEBUG
+#include <stdint.h>
+
 extern unsigned long long g_disable_isp_modules_en;
 extern unsigned long long g_disable_isp_modules_cfg_update;
 extern int g_disable_ispp_modules_en;
@@ -28,11 +30,11 @@ extern int g_bypass_exp_params;
 extern int g_bypass_isp_params;
 extern int g_bypass_ispp_params;
 extern int g_apply_init_params_only;
-extern int g_disable_algo_user_api_mask;
+extern uint64_t g_disable_algo_user_api_mask;
 extern void get_dbg_force_disable_mods_env();
 extern int get_rkaiq_runtime_dbg();
 #define CHECK_USER_API_ENABLE(mask) \
-    if (g_disable_algo_user_api_mask & (1 << mask)) { \
+    if (g_disable_algo_user_api_mask & (1ULL << mask)) { \
         ALOGE("algo module index %d user api disabled !", mask); \
         return XCAM_RETURN_NO_ERROR; \
     }
